+++ guillotina/api/files.py
@@ -4,6 +4,7 @@ from guillotina.api.content import DefaultOPTIONS
 from guillotina.api.service import DownloadService
 from guillotina.api.service import TraversableFieldService
 from guillotina.component import get_multi_adapter
+from guillotina.exceptions import FileNotFoundException
 from guillotina.interfaces import IAsyncBehavior
 from guillotina.interfaces import IFileManager
 from guillotina.interfaces import IResource
@@ -112,6 +113,34 @@ class UploadFile(TraversableFieldService):
 
 @configure.service(
     context=IResource,
+    method="DELETE",
+    permission="guillotina.DeleteContent",
+    name="@delete/{field_name}",
+    parameters=[{"in": "path", "name": "field_name", "required": True, "schema": {"type": "string"}}],
+    **_traversed_file_doc("Update the content of a file"),
+)
+@configure.service(
+    context=IResource,
+    method="DELETE",
+    permission="guillotina.DeleteContent",
+    name="@delete/{field_name}/{filename}",
+    parameters=[{"in": "path", "name": "field_name", "required": True, "schema": {"type": "string"}}],
+    **_traversed_file_doc("Update the content of a file"),
+)
+class DeleteFile(TraversableFieldService):
+    async def __call__(self):
+        if self.behavior is not None and IAsyncBehavior.implementedBy(self.behavior.__class__):
+            # providedBy not working here?
+            await self.behavior.load(create=True)
+        # We need to get the upload as async IO and look for an adapter
+        # for the field to save there by chunks
+        adapter = get_multi_adapter((self.context, self.request, self.field), IFileManager)
+        result = await adapter.delete()
+        return result
+
+
+@configure.service(
+    context=IResource,
     method="GET",
     permission="guillotina.ViewContent",
     name="@download/{field_name}",
@@ -140,7 +169,7 @@ class DownloadFile(TraversableFieldService):
         try:
             adapter = get_multi_adapter((self.context, self.request, self.field), IFileManager)
             return await self.handle(adapter, kwargs)
-        except AttributeError:
+        except (AttributeError, FileNotFoundException):
             # file does not exist
             return HTTPNotFound(content={"reason": "File does not exist"})